pyproject.toml
[project]
name = "qemzoo"
version = "0.1.0"
description = "The QEM Zoo: a catalog of quantum error mitigation techniques"
requires-python = ">=3.13"
dependencies = ["mitiq", "cirq-core"]
